% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{create_aoi}
\alias{create_aoi}
\title{create aoi}
\usage{
create_aoi(external_ref, geometry)
}
\arguments{
\item{external_ref}{Optional name for AOI}

\item{geometry}{geometry as list of lists or JSON}
}
\value{
AOIRecord
}
\description{
create aoi
}
