% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{create_subscription}
\alias{create_subscription}
\title{create subscription}
\usage{
create_subscription(aoi_id, dataset_id, external_ref = NULL)
}
\arguments{
\item{aoi_id}{ID for AOI to use for subscription}

\item{dataset_id}{ID for dataset to use for subscription}

\item{external_ref}{Optional name for subscription}
}
\value{
Subscription
}
\description{
create subscription
}
