% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_management.R
\name{create_user}
\alias{create_user}
\title{create user}
\usage{
create_user(first_name, last_name, email)
}
\arguments{
\item{first_name}{First name}

\item{last_name}{Last name}

\item{email}{Email}
}
\value{
User
}
\description{
create user
}
