% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DataRequest}
\alias{DataRequest}
\title{DataRequest Model}
\description{
DataRequest Model

DataRequest Model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataRequest-new}{\code{DataRequest$new()}}
\item \href{#method-DataRequest-as_list}{\code{DataRequest$as_list()}}
\item \href{#method-DataRequest-clone}{\code{DataRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequest$new(id, aoi_id, dataset_id, external_ref, created_at, created_by)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequest-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequest-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequest$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
