% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DataRequestCreate}
\alias{DataRequestCreate}
\title{DataRequestCreate Model}
\description{
DataRequestCreate Model

DataRequestCreate Model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataRequestCreate-new}{\code{DataRequestCreate$new()}}
\item \href{#method-DataRequestCreate-as_list}{\code{DataRequestCreate$as_list()}}
\item \href{#method-DataRequestCreate-clone}{\code{DataRequestCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequestCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequestCreate-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequestCreate$new(aoi_id, dataset_id, external_ref)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequestCreate-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequestCreate-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequestCreate$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequestCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequestCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequestCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
