% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DataRequestMetadata}
\alias{DataRequestMetadata}
\title{DataRequestMetadata Model}
\description{
DataRequestMetadata Model

DataRequestMetadata Model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataRequestMetadata-new}{\code{DataRequestMetadata$new()}}
\item \href{#method-DataRequestMetadata-as_list}{\code{DataRequestMetadata$as_list()}}
\item \href{#method-DataRequestMetadata-clone}{\code{DataRequestMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequestMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequestMetadata-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequestMetadata$new(
  provider_name,
  dataset_id,
  dataset_name,
  dataset_crs,
  aoi_id,
  data_request_id,
  files
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequestMetadata-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequestMetadata-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequestMetadata$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataRequestMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataRequestMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataRequestMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
