% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_management.R
\name{get_organisation_settings}
\alias{get_organisation_settings}
\title{get organisation settings}
\usage{
get_organisation_settings()
}
\value{
OrganisationSettings
}
\description{
get organisation settings
}
