% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_sf}
\alias{load_sf}
\title{}
\usage{
load_sf(subscription_id)
}
\arguments{
\item{subscription_id}{ID for subscription to load}
}
\value{
sf
}
\description{
all_bands <- list()
  urls <- metadata$files$url
  bands <- metadata$files$bands
}
\details{

}
