% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{OrganisationSettings}
\alias{OrganisationSettings}
\title{OrganisationSettings Model}
\description{
OrganisationSettings Model

OrganisationSettings Model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrganisationSettings-new}{\code{OrganisationSettings$new()}}
\item \href{#method-OrganisationSettings-as_list}{\code{OrganisationSettings$as_list()}}
\item \href{#method-OrganisationSettings-clone}{\code{OrganisationSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganisationSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrganisationSettings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationSettings$new(monthly_data_request_limit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganisationSettings-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-OrganisationSettings-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationSettings$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganisationSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrganisationSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganisationSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
