% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_management.R
\name{update_organisation_settings}
\alias{update_organisation_settings}
\title{Title}
\usage{
update_organisation_settings(monthly_data_request_limit = 50000)
}
\arguments{
\item{monthly_data_request_limit}{}
}
\description{
Title
}
