% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{User}
\alias{User}
\title{User Model}
\description{
User Model

User Model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-User-new}{\code{User$new()}}
\item \href{#method-User-as_list}{\code{User$as_list()}}
\item \href{#method-User-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-new"></a>}}
\if{latex}{\out{\hypertarget{method-User-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(id, first_name, last_name, email, created_at, created_by)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-User-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-clone"></a>}}
\if{latex}{\out{\hypertarget{method-User-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
